pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sm_blueprint_lib"
version = "0.0.15"
dependencies = [
  "numpy>=2.4.1", 
  "pygame>=2.6.1",
  "moderngl>=5.12.0", 
  "pyglm>=2.8.3", 
  "PyWavefront>=1.3.3"
]
authors = [
  { name="Maurice", email="mauriciotorrez00@gmail.com" },
  { name="lnventorsteve", email="dylanv60431@gmail.com" },
  { name="niknal357" },
]
description = "Scrap Mechanic Library for Blueprint manipulation."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]


[project.urls]
Homepage = "https://github.com/MauriceTZ/sm_blueprint_lib"
Issues = "https://github.com/MauriceTZ/sm_blueprint_lib/issues"